Territórios do saber - Agrupar componentes curriculares

+++ src/SME.SGP.Api/Controllers/GradeController.cs
@@ -19,7 +19,7 @@ namespace SME.SGP.Api.Controllers
         [ProducesResponseType(204)]
         [ProducesResponseType(typeof(IEnumerable<GradeComponenteTurmaAulasDto>), 200)]
         [ProducesResponseType(typeof(RetornoBaseDto), 500)]
-        public async Task<IActionResult> ObterGradeAulasTurma([FromQuery] DateTime data, string codigoTurma, int codigoDisciplina, [FromServices] IConsultasGrade consultasGrade)
+        public async Task<IActionResult> ObterGradeAulasTurma([FromQuery] DateTime data, string codigoTurma, long codigoDisciplina, [FromServices] IConsultasGrade consultasGrade)
         {
             var semana = (data.DayOfYear / 7) + 1;
             var horasGrade = await consultasGrade.ObterGradeAulasTurmaProfessor(codigoTurma, codigoDisciplina, semana.ToString(), data);
